migrate GatewayEvent to defguard_common crate
