+++ crates/defguard_event_router/Cargo.toml
@@ -9,6 +9,7 @@ rust-version.workspace = true
 
 [dependencies]
 # internal crates
+defguard_common = { workspace = true }
 defguard_core = { workspace = true }
 defguard_event_logger = { workspace = true }
 defguard_session_manager = { workspace = true }
@@ -19,4 +20,3 @@ tokio = { workspace = true }
 tracing = { workspace = true }
 
 [dev-dependencies]
-defguard_common = { workspace = true }
