+++ crates/defguard_event_router/src/lib.rs
@@ -19,10 +19,8 @@
 
 use std::sync::Arc;
 
-use defguard_core::{
-    events::{ApiEvent, BidiStreamEvent},
-    grpc::GatewayEvent,
-};
+use defguard_common::gateway_event::GatewayEvent;
+use defguard_core::events::{ApiEvent, BidiStreamEvent};
 use defguard_event_logger::message::{EventContext, EventLoggerMessage, LoggerEvent};
 use defguard_session_manager::events::SessionManagerEvent;
 use error::EventRouterError;
