+++ crates/defguard_gateway_manager/src/tests/gateway_manager/handler.rs
@@ -2,7 +2,7 @@
 mod support;
 
 use defguard_common::db::models::device::{DeviceInfo, WireguardNetworkDevice};
-use defguard_core::grpc::GatewayEvent;
+use defguard_common::gateway_event::GatewayEvent;
 use defguard_proto::gateway::{UpdateType, core_response};
 use sqlx::postgres::{PgConnectOptions, PgPoolOptions};
 use tonic::Status;
