+++ crates/defguard_proxy_manager/src/lib.rs
@@ -7,13 +7,13 @@ use std::{
 use std::{path::PathBuf, str::FromStr, sync::Mutex as StdMutex};
 
 use axum_extra::extract::cookie::Key;
+use defguard_common::gateway_event::GatewayEvent;
 use defguard_common::{
     db::{Id, models::proxy::Proxy},
     types::proxy::ProxyControlMessage,
 };
 use defguard_core::{
-    events::BidiStreamEvent,
-    grpc::{GatewayEvent, proxy::client_mfa::ClientLoginSession},
+    events::BidiStreamEvent, grpc::proxy::client_mfa::ClientLoginSession,
     version::IncompatibleComponents,
 };
 use defguard_proto::proxy::{CoreResponse, HttpsCerts, core_response};
