+++ crates/defguard_proxy_manager/src/tests/proxy_manager/handler/enrollment.rs
@@ -2,10 +2,8 @@ use defguard_common::db::models::{
     Device, Settings, User, biometric_auth::BiometricAuth, polling_token::PollingToken,
     settings::update_current_settings,
 };
-use defguard_core::{
-    events::{BidiStreamEventType, EnrollmentEvent},
-    grpc::GatewayEvent,
-};
+use defguard_common::gateway_event::GatewayEvent;
+use defguard_core::events::{BidiStreamEventType, EnrollmentEvent};
 use defguard_proto::{
     client_types::{ExistingDevice, MfaMethod, NewDevice, RegisterMobileAuthRequest},
     proxy::{CoreRequest, core_request, core_response},
